pyproject.toml
[project]
name = "cmdmark"
version = "0.1.0"
description = "A CLI tool for managing commands like bookmarks"
authors = [
    {name = "Po-wei Huang",email = "abc1199281@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["pyyaml (>=6.0.2,<7.0.0)"]

[tool.poetry.scripts]
cmdmark = "cmdmark.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
